requirements.txt
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
Brotli @ file:///work/perseverance-python-buildout/croot/brotli-split_1698805593785/work
causal-conv1d==1.2.0.post2
certifi @ file:///croot/certifi_1707229174982/work/certifi
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clearml==1.16.5
click==8.1.7
cloudpickle==3.0.0
comet-ml==3.47.1
configobj==5.0.9
contourpy==1.2.1
cos-python-sdk-v5==1.9.30
crcmod==1.7
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
dnspython==2.6.1
dulwich==0.22.3
einops==0.7.0
email_validator==2.2.0
et-xmlfile==1.1.0
everett==3.1.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
furl==2.1.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna @ file:///work/perseverance-python-buildout/croot/idna_1698845632828/work
imageio==2.34.0
iniconfig==2.0.0
Jinja2 @ file:///work/perseverance-python-buildout/croot/jinja2_1707343043683/work
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
lazy_loader==0.4
mamba-ssm==1.2.0.post1
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib==3.8.4
mdurl==0.1.2
MedPy==0.5.1
mkl-fft @ file:///work/perseverance-python-buildout/croot/mkl_fft_1698845673361/work
mkl-random @ file:///work/perseverance-python-buildout/croot/mkl_random_1698845720894/work
mkl-service==2.4.0
modelscope==1.16.1
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
multidict==6.0.5
multiprocess==0.70.16
networkx @ file:///work/perseverance-python-buildout/croot/networkx_1698865062738/work
ninja==1.11.1.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp312-cp312-linux_x86_64.whl#sha256=1d700f51d8b4fa684d858c9e3b56b1656bc5c82b6b79ff08d4e3b491c430059f
nvidia-cuda-runtime-cu12==12.6.77
onnx==1.17.0
onnxslim==0.1.36
opencv-python==4.10.0.84
openpyxl==3.1.2
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.1
pathlib2==2.3.7.post1
peewee==3.17.6
peft==0.12.0
pillow @ file:///croot/pillow_1707233021655/work
platformdirs==4.2.1
pluggy==1.4.0
protobuf==5.26.1
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.3
pyparsing==3.1.2
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytest==8.1.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytoolconfig==1.3.1
pytz==2024.1
PyYAML @ file:///work/perseverance-python-buildout/croot/pyyaml_1698849903511/work
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rope==1.13.0
rpds-py==0.20.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.17.0
setuptools==68.2.2
shellingham==1.5.4
SimpleITK==2.3.1
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
submitit==1.5.1
swanboard==0.1.3b2
swankit==0.1.1b1
swanlab==0.3.14
sympy @ file:///croot/sympy_1701397643339/work
tensorboardX==2.6.2.2
tensorrt==10.7.0
tensorrt-cu12==10.7.0
tensorrt-cu12-bindings==10.7.0
tensorrt-cu12-libs==10.7.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tiktoken==0.7.0
timm==0.9.16
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.39.3
triton==2.3.0
typer==0.12.3
typing_extensions @ file:///croot/typing_extensions_1705619912070/work
tzdata==2024.1
ujson==5.10.0
ultralytics==8.3.24
ultralytics-thop==2.0.10
urllib3 @ file:///croot/urllib3_1707770551213/work
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.41.2
wrapt==1.16.0
wurlitzer==3.1.1
xmltodict==0.13.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
